requirements.txt
matplotlib
numpy
pandas
diffusers==0.32.2
Pillow==12.0.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.43.0
accelerate==1.3.0
